pull out common functionality into a base class pt 1
