+++ lucene/core/src/java/org/apache/lucene/search/MultiTermQueryConstantScoreWrapper.java
@@ -16,59 +16,26 @@
  */
 package org.apache.lucene.search;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.PostingsEnum;
 import org.apache.lucene.index.Term;
-import org.apache.lucene.index.TermState;
 import org.apache.lucene.index.TermStates;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.index.TermsEnum;
 import org.apache.lucene.search.BooleanClause.Occur;
-import org.apache.lucene.util.Accountable;
-import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.DocIdSetBuilder;
-import org.apache.lucene.util.RamUsageEstimator;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 
 /**
  * This class also provides the functionality behind {@link MultiTermQuery#CONSTANT_SCORE_REWRITE}.
  * It tries to rewrite per-segment as a boolean query that returns a constant score and otherwise
  * fills a bit set with matches and builds a Scorer on top of this bit set.
  */
-final class MultiTermQueryConstantScoreWrapper<Q extends MultiTermQuery> extends Query
-    implements Accountable {
-
-  // mtq that matches 16 terms or less will be executed as a regular disjunction
-  private static final int BOOLEAN_REWRITE_TERM_COUNT_THRESHOLD = 16;
-
-  @Override
-  public long ramBytesUsed() {
-    if (query instanceof Accountable) {
-      return RamUsageEstimator.NUM_BYTES_OBJECT_HEADER
-          + RamUsageEstimator.NUM_BYTES_OBJECT_REF
-          + ((Accountable) query).ramBytesUsed();
-    }
-    return RamUsageEstimator.NUM_BYTES_OBJECT_HEADER
-        + RamUsageEstimator.NUM_BYTES_OBJECT_REF
-        + RamUsageEstimator.QUERY_DEFAULT_RAM_BYTES_USED;
-  }
-
-  private static class TermAndState {
-    final BytesRef term;
-    final TermState state;
-    final int docFreq;
-    final long totalTermFreq;
-
-    TermAndState(BytesRef term, TermState state, int docFreq, long totalTermFreq) {
-      this.term = term;
-      this.state = state;
-      this.docFreq = docFreq;
-      this.totalTermFreq = totalTermFreq;
-    }
-  }
+final class MultiTermQueryConstantScoreWrapper<Q extends MultiTermQuery> extends AbstractMultiTermQueryConstantScoreWrapper<Q> {
 
   private static class WeightOrDocIdSet {
     final Weight weight;
@@ -85,76 +52,14 @@ final class MultiTermQueryConstantScoreWrapper<Q extends MultiTermQuery> extends
     }
   }
 
-  protected final Q query;
-
-  /** Wrap a {@link MultiTermQuery} as a Filter. */
-  protected MultiTermQueryConstantScoreWrapper(Q query) {
-    this.query = query;
-  }
-
-  @Override
-  public String toString(String field) {
-    // query.toString should be ok for the filter, too, if the query boost is 1.0f
-    return query.toString(field);
-  }
-
-  @Override
-  public final boolean equals(final Object other) {
-    return sameClassAs(other)
-        && query.equals(((MultiTermQueryConstantScoreWrapper<?>) other).query);
-  }
-
-  @Override
-  public final int hashCode() {
-    return 31 * classHash() + query.hashCode();
-  }
-
-  /** Returns the encapsulated query */
-  public Q getQuery() {
-    return query;
-  }
-
-  /** Returns the field name for this query */
-  public final String getField() {
-    return query.getField();
+  MultiTermQueryConstantScoreWrapper(Q query) {
+    super(query);
   }
 
   @Override
   public Weight createWeight(IndexSearcher searcher, ScoreMode scoreMode, float boost)
       throws IOException {
-    return new ConstantScoreWeight(this, boost) {
-
-      /**
-       * Try to collect terms from the given terms enum and return true if all terms could be
-       * collected or if one of the iterated terms contains all docs for the field. If {@code false}
-       * is returned, the enum is left positioned on the next term.
-       */
-      private boolean collectTerms(int fieldDocCount, TermsEnum termsEnum, List<TermAndState> terms)
-          throws IOException {
-        final int threshold =
-            Math.min(BOOLEAN_REWRITE_TERM_COUNT_THRESHOLD, IndexSearcher.getMaxClauseCount());
-        for (int i = 0; i < threshold; ++i) {
-          final BytesRef term = termsEnum.next();
-          if (term == null) {
-            return true;
-          }
-          TermState state = termsEnum.termState();
-          int docFreq = termsEnum.docFreq();
-          TermAndState termAndState =
-              new TermAndState(
-                  BytesRef.deepCopyOf(term), state, docFreq, termsEnum.totalTermFreq());
-          if (fieldDocCount == docFreq) {
-            // If the term contains every document with a value for the field, we can ignore all
-            // other terms:
-            terms.clear();
-            terms.add(termAndState);
-            return true;
-          }
-          terms.add(termAndState);
-        }
-        return termsEnum.next() == null;
-      }
-
+    return new WrapperWeight(query, boost, scoreMode) {
       /**
        * On the given leaf context, try to either rewrite to a disjunction if there are few terms,
        * or build a bitset containing matching docs.
@@ -245,19 +150,6 @@ final class MultiTermQueryConstantScoreWrapper<Q extends MultiTermQuery> extends
       }
 
       @Override
-      public Matches matches(LeafReaderContext context, int doc) throws IOException {
-        final Terms terms = context.reader().terms(query.field);
-        if (terms == null) {
-          return null;
-        }
-        return MatchesUtils.forField(
-            query.field,
-            () ->
-                DisjunctionMatchesIterator.fromTermsEnum(
-                    context, doc, query, query.field, query.getTermsEnum(terms)));
-      }
-
-      @Override
       public Scorer scorer(LeafReaderContext context) throws IOException {
         final WeightOrDocIdSet weightOrBitSet = rewrite(context);
         if (weightOrBitSet.weight != null) {
@@ -266,18 +158,6 @@ final class MultiTermQueryConstantScoreWrapper<Q extends MultiTermQuery> extends
           return scorer(weightOrBitSet.set);
         }
       }
-
-      @Override
-      public boolean isCacheable(LeafReaderContext ctx) {
-        return true;
-      }
     };
-  }
-
-  @Override
-  public void visit(QueryVisitor visitor) {
-    if (visitor.acceptField(getField())) {
-      query.visit(visitor.getSubVisitor(Occur.FILTER, this));
-    }
   }
 }
